+++ package.json
@@ -2,6 +2,7 @@
   "name": "1-lesson",
   "version": "0.1.0",
   "private": true,
+  "proxy": "http://localhost:8081",
   "dependencies": {
     "@testing-library/jest-dom": "^6.2.0",
     "@testing-library/react": "^14.1.2",
@@ -12,8 +13,10 @@
     "@types/react-dom": "^18.2.18",
     "awesome-qr": "^2.1.5-rc.0",
     "axios": "^1.7.7",
+    "http-proxy-middleware": "^3.0.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-input-mask": "^2.0.4",
     "react-redux": "^9.1.2",
     "react-router-dom": "^6.28.0",
     "react-scripts": "5.0.1",
