+++ src/components/App/App.module.scss
@@ -15,4 +15,14 @@
   align-items: center;
   gap: 20px;
   overflow: hidden;
+  padding-bottom: 200px;
+}
+
+input[type="number"] {
+  -moz-appearance: textfield;
+}
+
+input[type="number"]::-webkit-inner-spin-button,
+input[type="number"]::-webkit-outer-spin-button {
+  -webkit-appearance: none;
 }
