+++ src/setupProxy.ts
@@ -1,0 +1,11 @@
+// import { createProxyMiddleware } from 'http-proxy-middleware';
+//
+// module.exports = function(app: any) {
+//   app.use(
+//     '/',
+//     createProxyMiddleware({
+//       target: 'http://localhost:8081/',
+//       changeOrigin: true,
+//     })
+//   );
+// };
